+++ ige/src/plugin/AnimationPlugin.cpp
@@ -19,9 +19,8 @@ using ige::core::App;
 using ige::ecs::EntityId;
 using ige::ecs::System;
 using ige::ecs::World;
-using ige::plugin::animation::Animation;
-using ige::plugin::animation::AnimationPlayer;
 using ige::plugin::animation::AnimationPlugin;
+using ige::plugin::animation::AnimationTrack;
 using ige::plugin::animation::Animator;
 using ige::plugin::animation::SkeletonPose;
 using ige::plugin::time::Time;
@@ -30,19 +29,12 @@ using std::optional;
 using std::chrono::duration;
 using std::chrono::duration_cast;
 
-SkeletonPose::SkeletonPose(Skeleton::Handle skeleton)
-    : skeleton(skeleton)
-    , joint_pose(skeleton->joints.size())
-    , global_pose(skeleton->joints.size())
-{
-}
-
 static void compute_joint_pose(
     std::size_t idx, std::size_t frame, SkeletonPose& pose,
-    const Animation& anim)
+    const AnimationTrack& track)
 {
     const auto& joint_info = pose.skeleton->joints[idx];
-    const auto& joint_anim = anim.clip->joints[idx];
+    const auto& joint_anim = track.clip->joints[idx];
 
     auto& joint_pose = pose.joint_pose[idx];
 
@@ -55,10 +47,10 @@ static void compute_joint_pose(
     auto& mat = pose.global_pose[idx];
 
     if (joint_info.parent) {
-        // perf: ideally, a child should always come *after* its parent so
-        // that its safe to assume the parent's transform is already computed
+        // TODO(perf): ideally, a child should always come *after* its parent
+        // so that its safe to assume the parent's transform is already computed
         // here
-        compute_joint_pose(*joint_info.parent, frame, pose, anim);
+        compute_joint_pose(*joint_info.parent, frame, pose, track);
 
         mat = pose.global_pose[*joint_info.parent];
     } else {
@@ -70,7 +62,47 @@ static void compute_joint_pose(
     mat = glm::scale(mat, joint_pose.scale);
 }
 
-namespace systems {
+template <typename Duration>
+static void
+update_animation_track(World& world, AnimationTrack& track, Duration delta)
+{
+    auto pose = world.get_component<SkeletonPose>(track.target);
+
+    if (!pose) {
+        return;
+    }
+
+    track.current_time
+        += duration_cast<AnimationTrack::Duration>(track.playback_rate * delta);
+
+    if (track.loop) {
+        track.current_time %= track.clip->duration;
+    }
+
+    // the current animation frame
+    const std::size_t frame = static_cast<std::size_t>(
+        track.current_time / track.clip->sample_duration);
+
+    // how many joints to animate
+    // should be equal to the number of joints of the skeleton
+    std::size_t joint_count = track.clip->joints.size();
+
+    // TODO: maybe wrap this in some ifdef DEBUG?
+    if (joint_count != pose->skeleton->joints.size()) {
+        std::cerr << "[WARN] Number of joints in the animation (" << joint_count
+                  << ") doesn't match the number of joints in the skeleton ("
+                  << pose->skeleton->joints.size() << ") - skipping animation."
+                  << std::endl;
+        return;
+    }
+
+    pose->joint_pose.resize(joint_count);
+    pose->global_pose.resize(joint_count);
+
+    for (std::size_t j = 0; j < joint_count; j++) {
+        compute_joint_pose(j, frame, *pose, track);
+    }
+}
 
 static void update_animations(World& world)
 {
@@ -84,59 +116,21 @@ static void update_animations(World& world)
     }
 
     for (auto [ent, animator] : world.query<Animator>()) {
-        if (!animator.player) {
-            continue;
-        }
-
-        const auto& anim = animator.animations[animator.player->index];
-        auto pose = world.get_component<SkeletonPose>(anim.target);
-
-        if (!pose) {
-            continue;
-        }
-
-        if (!animator.player->paused) {
-            animator.player->current_time += time->delta();
-
-            if (animator.player->loop) {
-                animator.player->current_time %= anim.clip->duration;
+        for (auto& track : animator.tracks()) {
+            // skip every track that won't participate in the calculation of the
+            // final pose
+            if (track.weight != 1.0f) {
+                // TODO: animation blending!
+                continue;
             }
-        }
 
-        // the current animation frame
-        std::size_t frame
-            = animator.player->current_time / anim.clip->sample_duration;
-
-        // how many joints to animate
-        // should be = how many joints in the skeleton
-        std::size_t joint_count = anim.clip->joints.size();
-
-        if (joint_count != pose->skeleton->joints.size()) {
-            std::cerr
-                << "[WARN] Number of joints in the animation (" << joint_count
-                << ") doesn't match the number of joints in the skeleton ("
-                << pose->skeleton->joints.size() << ") - skipping animation."
-                << std::endl;
-            continue;
-        }
-
-        if (pose->joint_pose.size() != joint_count) {
-            pose->joint_pose.resize(joint_count);
-        }
-
-        if (pose->global_pose.size() != joint_count) {
-            pose->global_pose.resize(joint_count);
-        }
-
-        for (std::size_t j = 0; j < joint_count; j++) {
-            compute_joint_pose(j, frame, *pose, anim);
+            update_animation_track(
+                world, track, animator.playback_rate * time->delta());
         }
     }
 }
 
-}
-
 void AnimationPlugin::plug(App::Builder& builder) const
 {
-    builder.add_system(System::from(systems::update_animations));
+    builder.add_system(System::from(update_animations));
 }